Move program main code to mod system
